+++ src/Components/UserProfile/UserProfile.js
@@ -5,9 +5,12 @@ import Teacher from "../../Teacherdata";
 import TeacherCom from "../../Components/UserProfile/Teacher";
 import Student from "../../Components/UserProfile/Student";
 export default function UserProfile() {
+
   const [teacherData, setTeacherData] = useState([]);
+
   const userId = useLocation();
   const id = userId.state;
+  
   function isEmailExists(id) {
     for (let i = 0; i < Teacher.length; i++) {
       if (Teacher[i].id === id) {