+++ changelog.md
@@ -1,3 +1,4 @@
  # Added
  - Create User Role , Login by admin , Protected Routes [making-principal-admin]
  - Create UI of student and teacher module. fixed issues in login page [teacher-module]
+ - Routing of UserProfile of student and teacher [userprofile]